Forced cast to prevent MSVC compiler warning

+++ library/ssl_cli.c
@@ -1071,7 +1071,7 @@ static int ssl_write_client_key_exchange( ssl_context *ssl )
         i = 6;
 
         ret = dhm_make_public( &ssl->handshake->dhm_ctx,
-                                mpi_size( &ssl->handshake->dhm_ctx.P ),
+                               (int) mpi_size( &ssl->handshake->dhm_ctx.P ),
                                &ssl->out_msg[i], n,
                                 ssl->f_rng, ssl->p_rng );
         if( ret != 0 )
